Move InlineErrorType to types
